testing char create
